+++ exonum/src/blockchain/mod.rs
@@ -278,7 +278,7 @@ impl Blockchain {
             for service in self.service_map.values() {
                 // Skip execution for genesis block.
                 if height > Height(0) {
-                    service_execute_before_commit(service.as_ref(), &mut fork);
+                    service_before_commit(service.as_ref(), &mut fork);
                 }
             }
 
@@ -552,10 +552,10 @@ impl Blockchain {
     }
 }
 
-fn service_execute_before_commit(service: &Service, fork: &mut Fork) {
+fn service_before_commit(service: &Service, fork: &mut Fork) {
     fork.checkpoint();
     match panic::catch_unwind(panic::AssertUnwindSafe(|| {
-        service.execute_before_commit(fork)
+        service.before_commit(fork)
     })) {
         Ok(..) => fork.commit(),
         Err(err) => {
@@ -565,7 +565,7 @@ fn service_execute_before_commit(service: &Service, fork: &mut Fork) {
             }
             fork.rollback();
             error!(
-                "{} service execute_before_commit failed with error: {:?}",
+                "{} service before_commit failed with error: {:?}",
                 service.service_name(),
                 err
             );