+++ CHANGELOG.md
@@ -46,7 +46,9 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 - `Field`, `CryptoHash`, `StorageValue` and `ExonumJson` traits have been
   implemented for `chrono::Duration` structure. (#653)
 
-- `execute_before_commit` method has been added in `Service` trait. (#667) (#715)
+- `before_commit` method has been added in `Service` trait. (#667) (#715)
+
+- `handle_commit` method in `Service` trait has been has been renamed to `on_commit`. (#715)
 
 - `Field`, `CryptoHash`, `StorageKey`, `StorageValue` and `ExonumJson` traits
   have been implemented for `rust_decimal::Decimal`. (#671)